+++ Binaries/dcmtk-source/dcmdata/dcvrfd.cc
@@ -122,7 +122,7 @@ void DcmFloatingPointDouble::print(ostream &out,
                     OFStandard::ftoa(buffer + 1, sizeof(buffer) - 1, *doubleVals);
                 }
                 /* check whether current value sticks to the length limit */
-                newLength = printedLength + strlen(buffer);
+                newLength = printedLength + (unsigned int)strlen(buffer);
                 if ((newLength <= maxLength) && ((i + 1 == count) || (newLength + 3 <= maxLength)))
                 {
                     out << buffer;
