+++ DCM Framework/jpegdatasrc.m
@@ -132,7 +132,7 @@ fill_input_buffer (j_decompress_ptr cinfo)
 	
   data_src_ptr src = (data_src_ptr) cinfo->src;
   size_t nbytes;
-	nbytes = readFromData(src->data, src->buffer, src->currentPosition, INPUT_BUF_SIZE);
+	nbytes = readFromData(src->data, src->buffer, (int)src->currentPosition, INPUT_BUF_SIZE);
   //nbytes = JFREAD(src->infile, src->buffer, INPUT_BUF_SIZE);
 
   if (nbytes <= 0) {
